third_party/WebKit/LayoutTests/crypto/digest.html

<!DOCTYPE html>
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Tests cypto.subtle.digest.");

jsTestIsAsync = true;

// Each sub-test run in this file is asynchronous. Chaining them together
// manually leads to very unreadable code, due to having closures within
// closures within closures. Instead of doing that, each subtest calls
// "startNextTest()" once it has completed.

currentTestIndex = 0;

function startNextTest()
{
    var currentTest = allTests[currentTestIndex++];

    if (!currentTest) {
        finishJSTest();
        return;
    }

    currentTest();
}

function rejectHandler(value, justPrint)
{
    debug("    rejected with value of " + value);
    if (!justPrint)
        startNextTest();
}

function resultHandler(buffer, justPrint)
{
    debug("    = " + byteArrayToHexString(new Uint8Array(buffer)));
    if (!justPrint)
        startNextTest();
}

function failHandler(value)
{
    testFailed(value);
    startNextTest();
}

allTests = [
    function()
    {
        debug("SHA1 of []")
        crypto.subtle.digest({name: 'sha-1'}, new Uint8Array([])).then(resultHandler, rejectHandler);
    },

    function()
    {
        debug("SHA1 of [0x0]")
        crypto.subtle.digest({name: 'sha-1'}, new Uint8Array([0])).then(resultHandler, rejectHandler);
    },

    function()
    {
        debug("SHA-256 rejects (dummy implementation)");
        crypto.subtle.digest({name: 'sha-256'}, new Uint8Array([])).then(resultHandler, rejectHandler);
    },

    function()
    {
        debug("Error (dummy implementation rejects this input)");
        var data = new Uint8Array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);
        crypto.subtle.digest({name: 'sha-1'}, data).then(resultHandler, rejectHandler);
    },

    function()
    {
        shouldThrow("crypto.subtle.digest({name: 'sha-1'})");
        shouldThrow("crypto.subtle.digest({name: 'sha-1'}, null)");
        shouldThrow("crypto.subtle.digest({name: 'sha-1'}, 10)");
        startNextTest();
    },
];

// Begin!
startNextTest();

</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
